[ne2k][eth] major cleanup, more inline documentation, small driver fixes.
