--- elks/arch/i86/drivers/net/build.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-
-# Build test program for target
-# within a single segment CS=DS=ES=SS
-
-as86 -0 -o ne2k-low.o ne2k-low.s
-as86 -0 -o ne2k-phy.o ne2k-phy.s
-as86 -0 -o ne2k-mac.o ne2k-mac.s
-
-bcc -ansi -0 -c -o ne2k-test.o ne2k-test.c
-
-ld86 -0 -d -M -o ne2k.bin ne2k-low.o ne2k-phy.o ne2k-mac.o ne2k-test.o | sort -k4 > ne2k.map
